dashboard/requirements.txt
Flask
Flask-CORS
SQLAlchemy
psycopg2-binary
python-dotenv